llamaindex/agents/basic/requirements.txt
llama-index
llama-index-llms-ollama 
llama-index-callbacks-arize-phoenix
python-dotenv
